+++ src/ListBroking/AppBundle/Engine/Filter/ContactFilter/BasicContactFilter.php
@@ -9,6 +9,7 @@
 namespace ListBroking\AppBundle\Engine\Filter\ContactFilter;
 
 use Doctrine\ORM\Query\Expr\Andx;
+use Doctrine\ORM\Query\Expr\Composite;
 use Doctrine\ORM\QueryBuilder;
 use ListBroking\AppBundle\Engine\Filter\ContactFilterInterface;
 use ListBroking\AppBundle\Form\FiltersType;
@@ -40,13 +41,13 @@ class BasicContactFilter implements ContactFilterInterface
                     switch ( $filter['filter_operation'] )
                     {
                         case FiltersType::INCLUSION_FILTER:
-                            $exp_bucket[$filter['field']][$filter['opt']][$filter['filter_operation']]['expression'] = $qb->expr()
-                                                                                                                          ->in($filter['field'], ':' . $name)
+                            $exp_bucket[$filter['field']][$filter['opt']][$filter['filter_operation']]['expressions'] = $qb->expr()
+                                                                                                                           ->in($filter['field'], ':' . $name)
                             ;
                             break;
                         case FiltersType::EXCLUSION_FILTER:
-                            $exp_bucket[$filter['field']][$filter['opt']][$filter['filter_operation']]['expression'] = $qb->expr()
-                                                                                                                          ->notIn($filter['field'], ':' . $name)
+                            $exp_bucket[$filter['field']][$filter['opt']][$filter['filter_operation']]['expressions'] = $qb->expr()
+                                                                                                                           ->notIn($filter['field'], ':' . $name)
                             ;
                             break;
                         default:
@@ -57,32 +58,35 @@ class BasicContactFilter implements ContactFilterInterface
                     break;
                 case FiltersType::BETWEEN_OPERATION:
 
-                    $name_x = str_replace('.', '_', sprintf("between_filter_x_%s_%s", $filter['filter_operation'], $filter['field']));
-                    $name_y = str_replace('.', '_', sprintf("between_filter_y_%s_%s", $filter['filter_operation'], $filter['field']));
-
-                    // Inclusion or Exclusion Filter
-                    switch ( $filter['filter_operation'] )
+                    foreach ( $filter['value'] as $key => $value )
                     {
-                        case FiltersType::INCLUSION_FILTER:
-                            $exp_bucket[$filter['field']][$filter['opt']][$filter['filter_operation']]['expression'] = $qb->expr()
-                                                                                                                          ->between($filter['field'], ":" . $name_x, ":" . $name_y)
-                            ;
-                            break;
-                        case FiltersType::EXCLUSION_FILTER:
-                            $exp_bucket[$filter['field']][$filter['opt']][$filter['filter_operation']]['expression'] = $qb->expr()
-                                                                                                                          ->not($qb->expr()
-                                                                                                                                   ->between($filter['field'], ":" . $name_x, ":" . $name_y))
-                            ;
-                            break;
-                        default:
-                            break;
+                        $name_x = str_replace('.', '_', sprintf("between_filter_x_%s_%s_%s", $filter['filter_operation'], $filter['field'], $key));
+                        $name_y = str_replace('.', '_', sprintf("between_filter_y_%s_%s_%s", $filter['filter_operation'], $filter['field'], $key));
+
+                        // Inclusion or Exclusion Filter
+                        switch ( $filter['filter_operation'] )
+                        {
+                            case FiltersType::INCLUSION_FILTER:
+                                $exp_bucket[$filter['field']][$filter['opt']][$filter['filter_operation']]['expressions'][] = $qb->expr()
+                                                                                                                                 ->between($filter['field'], ":" . $name_x, ":" . $name_y)
+                                ;
+                                break;
+                            case FiltersType::EXCLUSION_FILTER:
+                                $exp_bucket[$filter['field']][$filter['opt']][$filter['filter_operation']]['expressions'][] = $qb->expr()
+                                                                                                                                 ->not($qb->expr()
+                                                                                                                                          ->between($filter['field'], ":" . $name_x, ":" . $name_y))
+                                ;
+                                break;
+                            default:
+                                break;
+                        }
+
+                        $exp_bucket[$filter['field']][$filter['opt']][$filter['filter_operation']]['parameter_id'][] = $name_x;
+                        $exp_bucket[$filter['field']][$filter['opt']][$filter['filter_operation']]['parameter_id'][] = $name_y;
+
+                        $exp_bucket[$filter['field']][$filter['opt']][$filter['filter_operation']]['values'][$name_x] = $value[0];
+                        $exp_bucket[$filter['field']][$filter['opt']][$filter['filter_operation']]['values'][$name_y] = $value[1];
                     }
-
-                    $exp_bucket[$filter['field']][$filter['opt']][$filter['filter_operation']]['parameter_id']['x'] = $name_x;
-                    $exp_bucket[$filter['field']][$filter['opt']][$filter['filter_operation']]['parameter_id']['y'] = $name_y;
-
-                    $exp_bucket[$filter['field']][$filter['opt']][$filter['filter_operation']]['values']['x'] = $filter['value'][0];
-                    $exp_bucket[$filter['field']][$filter['opt']][$filter['filter_operation']]['values']['y'] = $filter['value'][1];
                     break;
             }
         }
@@ -102,29 +106,57 @@ class BasicContactFilter implements ContactFilterInterface
             foreach ( $operations as $operation => $filter_operations )
             {
                 /** @var AndX $expressions */
-                $expressions = $qb->expr()
-                                  ->andX()
+                $expr = $qb->expr()
+                           ->andX()
                 ;
 
                 if ( array_key_exists('inclusion', $filter_operations) )
                 {
                     $inclusion_filter = $filter_operations['inclusion'];
-                    $expressions->add($inclusion_filter['expression']);
+                    $inclusion_or = $qb->expr()
+                                       ->orX()
+                    ;
 
+                    $this->addExpressions($expr, $inclusion_or, $inclusion_filter['expressions']);
                     $this->addParameters($qb, $operation, $inclusion_filter['parameter_id'], $inclusion_filter['values']);
                 }
 
                 if ( array_key_exists('exclusion', $filter_operations) )
                 {
                     $exclusion_filter = $filter_operations['exclusion'];
-                    $expressions->add($exclusion_filter['expression']);
+                    $exclusion_and = $qb->expr()
+                                        ->andX()
+                    ;
 
+                    $this->addExpressions($expr, $exclusion_and, $exclusion_filter['expressions']);
                     $this->addParameters($qb, $operation, $exclusion_filter['parameter_id'], $exclusion_filter['values']);
                 }
-                $orX->add($expressions);
+                $orX->add($expr);
             }
             $andX->add($orX);
         }
+    }
+
+    /**
+     * Add the given parts to the QueryBuilder
+     *
+     * @param Composite    $expr
+     * @param Composite    $comp
+     * @param              $parts
+     */
+    private function addExpressions (Composite $expr, Composite $comp, $parts)
+    {
+        if ( is_array($parts) )
+        {
+            foreach ( $parts as $part )
+            {
+                $comp->add($part);
+            }
+            $expr->add($comp);
+
+            return;
+        }
+        $expr->add($parts);
     }
 
     /**
@@ -132,7 +164,7 @@ class BasicContactFilter implements ContactFilterInterface
      *
      * @param QueryBuilder $qb
      * @param string       $operation
-     * @param string       $parameter_id
+     * @param string|array $parameter_id
      * @param array        $values
      */
     private function addParameters (QueryBuilder $qb, $operation, $parameter_id, $values)
@@ -141,10 +173,11 @@ class BasicContactFilter implements ContactFilterInterface
         {
             case FiltersType::EQUAL_OPERATION:
 
-                if(count($values) == 1)
+                if ( count($values) == 1 )
                 {
                     $qb->setParameter($parameter_id, $values[0]);
-                }else
+                }
+                else
                 {
                     $final_values = array();
                     foreach ( $values as $value )
@@ -155,8 +188,10 @@ class BasicContactFilter implements ContactFilterInterface
                 }
                 break;
             case FiltersType::BETWEEN_OPERATION:
-                $qb->setParameter($parameter_id['x'], $values['x']);
-                $qb->setParameter($parameter_id['y'], $values['y']);
+                foreach ( $parameter_id as $id )
+                {
+                    $qb->setParameter($id, $values[$id]);
+                }
                 break;
         }
     }